I added the two micropython scripts I have been using - will updated readme with info on that eventually. ESP32_Script.py is the one I have mainly been using for testing purposes and is what I ill hope to use as a final product (its much more clean). Original is the script we created back for the client showcase in December (its a bit of a mess).
In main.py I updated some small things -
you can now have the app running and if the microcontroller program isn't running yet thats fine, the app will find it once its on.

ESP32_Script.py now connects and allows for data to be sent to the microcontroller - only issue is that the data always sends as the same thing regardless of what you change it to on the app. The app sending wise is fine so the issue seems to be in the script or in the transmission itself. I am leaning towards transmissio.